jethexa/requirements.txt
firebase-admin
requests
numpy
Pillow
Adafruit-SSD1306
Jetson.GPIO
